pyproject.toml
[project]
name = "celltype-refinery"
version = "1.0.0"
description = "Tissue-agnostic cell-type annotation and refinement for spatial proteomics data"
authors = [{name = "CellType-Refinery Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = [
    "cell-type annotation",
    "spatial proteomics",
    "single-cell analysis",
    "CODEX",
    "PhenoCycler",
    "multiplexed imaging",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "anndata>=0.10",
    "scanpy>=1.9",
    "leidenalg>=0.10",
    "pyyaml>=6.0",
    "click>=8.0",
    "tqdm>=4.65",
    "statsmodels>=0.14",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
]
gpu = [
    "rapids-singlecell>=0.10",
    "cupy-cuda12x>=12.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
]
all = [
    "celltype-refinery[viz,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/celltype-refinery/celltype-refinery"
Documentation = "https://celltype-refinery.readthedocs.io"
Repository = "https://github.com/celltype-refinery/celltype-refinery"
Issues = "https://github.com/celltype-refinery/celltype-refinery/issues"

[project.scripts]
celltype-refinery = "celltype_refinery.cli.main:main"

[project.entry-points."celltype_refinery.tissues"]
# Tissue-specific plugins can register here
# example: fallopian_tube = "celltype_refinery.tissues.fallopian_tube:FallopianTubeConfig"

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "celltype_refinery/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/celltype_refinery",
    "/configs",
    "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["celltype_refinery"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["celltype_refinery"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["tests/", "docs/"]

[tool.coverage.run]
source = ["celltype_refinery"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
